src/third_party/web_platform_tests/websockets/interfaces/WebSocket/send/005.html

<!doctype html>
<title>WebSockets: send() return value</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onopen = t.step_func(function(e) {
    // SpiderMonkey erroneously does not resolve a void function to "undefined".
    //assert_equals(ws.send('test'), undefined);
    ws.send('test');
    t.done();
  });
});
</script>
